backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
yfinance
pandas
numpy
sqlalchemy
psycopg2-binary
apscheduler
resend
python-dotenv
email-validator
aiofiles
httpx
